+++ internal/buildinfo/buildinfo.go
@@ -2,8 +2,8 @@ package buildinfo
 
 const (
 	Name       = "Simple Bridge Panel"
-	Version    = "1.5.1"
-	Prerelease = false
+	Version    = "1.5.2"
+	Prerelease = true
 	Repository = "silenceremember/sbp-panel"
 )